move workflow sub-components to workflow folder
