+++ frontend/src/app/components/camp-workflow/camp-basics/camp-basics.component.html
