+++ frontend/src/app/components/camp-workflow/camp-book/camp-book.component.html
