+++ frontend/src/app/components/camp-workflow/camp-workflow.component.ts
@@ -2,10 +2,10 @@ import {Component, Input} from '@angular/core';
 import {Observable} from 'rxjs/Observable';
 import {Camp, Hotel} from '../../model/backend-typings';
 import {MaterializeDirective} from 'angular2-materialize/dist/index';
-import {CampBasicsComponent} from '../camp-basics/camp-basics.component';
-import {CampDateComponent} from '../camp-date/camp-date.component';
-import {CampBookComponent} from '../camp-book/camp-book.component';
-import {CampReviewComponent} from '../camp-review/camp-review.component';
+import {CampBasicsComponent} from './camp-basics/camp-basics.component';
+import {CampDateComponent} from './camp-date/camp-date.component';
+import {CampBookComponent} from './camp-book/camp-book.component';
+import {CampReviewComponent} from './camp-review/camp-review.component';
 import {HotelOverviewComponent} from '../hotel-overview/hotel-overview.component';
 
 @Component({
